server/requirements.txt
Flask
flask-cors
langchain
openai
pandas
python-dotenv
requests
spotipy
tqdm